requirements_colab.txt
tqdm==4.64.0
pydantic==1.9.1
plotly==5.9.0
wandb==0.12.21
pandas
pytorch-lightning==1.6.5
avalanche-lib==0.2.1
torchvision==0.13.0
torchtext==0.13.0